Move sdk section component
